Move read-path sanitization into the entries service (#956)
The sanitized-content resolution (resolveSanitizedContent / selectFullEntry /
toFullEntry) lived in the entries tRPC router, so only entries.get and
entries.fetchFullContent served sanitized HTML — MCP get_entry, Google Reader,
and Wallabag read via services getEntry/getEntries and served raw feed HTML.

Move the chokepoint into src/server/services/entries.ts and switch
getEntry/getEntries to select the persisted *_sanitized columns and resolve
them (with the same version-gated self-heal), so all four consumers get the
sanitization guarantee. The router now imports selectFullEntry/toFullEntry
from the service.
