+++ CLAUDE.md
@@ -137,8 +137,8 @@ Prefer SAX-style parsing unless the algorithm requires a DOM.
 
 ### Sanitizing untrusted HTML
 
-Entry bodies (and AI summaries) are rendered via `dangerouslySetInnerHTML`, so untrusted HTML is sanitized **on the server, on the read path** with `sanitizeEntryHtml` from `@/server/html/sanitize` (a `sanitize-html` wrapper — pure Node, no DOM/jsdom). The client renders trusted HTML and ships no sanitizer. The chokepoints are `toFullEntry` in the entries router (covers `entries.get` + `entries.fetchFullContent`) and summary generation in `summarization.ts`. Do **not** reintroduce a client-side sanitizer — the old `isomorphic-dompurify` pulled `jsdom` into the production server bundle. Feed `summary`/`title`/`author` are rendered as escaped text, not HTML; keep them that way.
+Entry bodies (and AI summaries) are rendered via `dangerouslySetInnerHTML`, so untrusted HTML is sanitized **on the server, on the read path** with `sanitizeEntryHtml` from `@/server/html/sanitize` (a `sanitize-html` wrapper — pure Node, no DOM/jsdom). The client renders trusted HTML and ships no sanitizer. The chokepoints live in the **services layer**: `toFullEntry`/`getEntry`/`getEntries` in `src/server/services/entries.ts` (covering tRPC `entries.get` + `entries.fetchFullContent`, MCP `get_entry`, Google Reader, and Wallabag) and summary generation in `summarization.ts`. Do **not** reintroduce a client-side sanitizer — the old `isomorphic-dompurify` pulled `jsdom` into the production server bundle. Feed `summary`/`title`/`author` are rendered as escaped text, not HTML; keep them that way.
 
 `sanitizeEntryHtml` also runs a pre-sanitization transform, `convertMathJaxChtmlToMathml` from `@/server/html/mathjax-chtml`: some sources (notably LessWrong) deliver math pre-rendered as MathJax v3 CHTML (`<mjx-*>` elements + a `<style>` block, where each glyph is an empty `<mjx-c class="mjx-c1D465 …">` whose character lives in CSS). The sanitizer drops `<style>` and unknown tags, so that math would otherwise vanish entirely; the transform rewrites the CHTML tree to presentation MathML (already on the allow-list, renders natively) before sanitization. It's a cheap no-op (string check) for the common case with no embedded math.
 
-Sanitizing a large body is the dominant cost of `entries.get` (~50ms per ~700KB, ×4 fields), so sanitized output is **persisted in the database** (`entries.content_*_sanitized` / `full_content_*_sanitized`, each stamped with `content_sanitized_version` / `full_content_sanitized_version`). Every write path funnels its insert/update through `withSanitizedEntryContent` from `@/server/html/sanitize-entry` — the single place that derives the sanitized columns from raw content — so the invariant can't be forgotten at a new write site. The read path (`resolveSanitizedContent` in the entries router) serves the stored value when its version matches `SANITIZER_VERSION`, and otherwise re-sanitizes from the raw columns and self-heals (fire-and-forget persist). **Bump `SANITIZER_VERSION` in `sanitize.ts` whenever `SANITIZE_OPTIONS` _or_ the pre-sanitization transforms (`convertMathJaxChtmlToMathml`) change** — that marks every row stale so it is re-sanitized on next read (raw content is retained precisely so this works without a re-fetch).
+Sanitizing a large body is the dominant cost of `entries.get` (~50ms per ~700KB, ×4 fields), so sanitized output is **persisted in the database** (`entries.content_*_sanitized` / `full_content_*_sanitized`, each stamped with `content_sanitized_version` / `full_content_sanitized_version`). Every write path funnels its insert/update through `withSanitizedEntryContent` from `@/server/html/sanitize-entry` — the single place that derives the sanitized columns from raw content — so the invariant can't be forgotten at a new write site. The read path (`resolveSanitizedContent` in `src/server/services/entries.ts`) serves the stored value when its version matches `SANITIZER_VERSION`, and otherwise re-sanitizes from the raw columns and self-heals (fire-and-forget persist). **Bump `SANITIZER_VERSION` in `sanitize.ts` whenever `SANITIZE_OPTIONS` _or_ the pre-sanitization transforms (`convertMathJaxChtmlToMathml`) change** — that marks every row stale so it is re-sanitized on next read (raw content is retained precisely so this works without a re-fetch).
