+++ src/server/trpc/routers/entries.ts
@@ -10,7 +10,7 @@
  */
 
 import { z } from "zod";
-import { eq, and, sql } from "drizzle-orm";
+import { eq, and } from "drizzle-orm";
 import { createHash } from "crypto";
 
 import {
@@ -21,19 +21,11 @@ import {
 import { API_TOKEN_SCOPES } from "@/server/auth/api-token";
 import { errors } from "../errors";
 import { uuidSchema } from "../validation";
-import {
-  entries,
-  feeds,
-  subscriptions,
-  tags,
-  visibleEntries,
-  narrationContent,
-} from "@/server/db/schema";
+import { entries, tags, narrationContent } from "@/server/db/schema";
 import { fetchFullContent as fetchFullContentFromUrl } from "@/server/services/full-content";
 import * as entriesService from "@/server/services/entries";
 import * as countsService from "@/server/services/counts";
 import { getSubscriptionFeedIds } from "@/server/services/entry-filters";
-import { sanitizeEntryHtml, SANITIZER_VERSION } from "@/server/html/sanitize";
 import { withSanitizedEntryContent } from "@/server/html/sanitize-entry";
 import { publishEntryStateChanged } from "@/server/redis/pubsub";
 import { logger } from "@/lib/logger";
@@ -218,240 +210,13 @@ const setStarredOutputSchema = z.object({
 });
 
 // ============================================================================
-// Shared Select Fields
-// ============================================================================
-
-/**
- * Fields to select when fetching a full entry from visibleEntries + feeds + subscriptions.
- * Used by both `get` and `fetchFullContent` to avoid duplicating the 20+ field list.
- */
-const fullEntrySelectFields = {
-  id: visibleEntries.id,
-  feedId: visibleEntries.feedId,
-  type: visibleEntries.type,
-  url: visibleEntries.url,
-  title: visibleEntries.title,
-  author: visibleEntries.author,
-  // Sanitized content is served to the client; the matching version columns let
-  // the read path detect when the allow-list changed and re-sanitize from raw.
-  contentOriginalSanitized: visibleEntries.contentOriginalSanitized,
-  contentCleanedSanitized: visibleEntries.contentCleanedSanitized,
-  contentSanitizedVersion: visibleEntries.contentSanitizedVersion,
-  summary: visibleEntries.summary,
-  publishedAt: visibleEntries.publishedAt,
-  fetchedAt: visibleEntries.fetchedAt,
-  read: visibleEntries.read,
-  starred: visibleEntries.starred,
-  updatedAt: visibleEntries.updatedAt,
-  subscriptionId: visibleEntries.subscriptionId,
-  siteName: visibleEntries.siteName,
-  feedTitle: feeds.title,
-  feedUrl: feeds.url,
-  unsubscribeUrl: visibleEntries.unsubscribeUrl,
-  fullContentOriginalSanitized: visibleEntries.fullContentOriginalSanitized,
-  fullContentCleanedSanitized: visibleEntries.fullContentCleanedSanitized,
-  fullContentSanitizedVersion: visibleEntries.fullContentSanitizedVersion,
-  fullContentFetchedAt: visibleEntries.fullContentFetchedAt,
-  fullContentError: visibleEntries.fullContentError,
-  contentHash: visibleEntries.contentHash,
-  hasMarkedReadOnList: visibleEntries.hasMarkedReadOnList,
-  hasMarkedUnread: visibleEntries.hasMarkedUnread,
-  hasStarred: visibleEntries.hasStarred,
-  fetchFullContent: subscriptions.fetchFullContent,
-};
-
-/**
- * Fetch a single full entry by ID for a user.
- * Queries visibleEntries joined with feeds and subscriptions.
- * Returns null if the entry is not found or not visible to the user.
- */
-async function selectFullEntry(
-  db: typeof import("@/server/db").db,
-  userId: string,
-  entryId: string
-) {
-  const result = await db
-    .select(fullEntrySelectFields)
-    .from(visibleEntries)
-    .innerJoin(feeds, eq(visibleEntries.feedId, feeds.id))
-    .leftJoin(subscriptions, eq(visibleEntries.subscriptionId, subscriptions.id))
-    .where(and(eq(visibleEntries.id, entryId), eq(visibleEntries.userId, userId)))
-    .limit(1);
-
-  return result.length > 0 ? result[0] : null;
-}
-
-/**
- * Resolve an entry's sanitized content for display.
- *
- * Entry bodies come from untrusted feeds and are rendered via
- * `dangerouslySetInnerHTML`, so they must be sanitized. Sanitization is
- * persisted in the `*_sanitized` columns at write time (see
- * `@/server/html/sanitize-entry`), so the common case is a pure read of the
- * stored values. When the stored version doesn't match the current
- * `SANITIZER_VERSION` — pre-migration rows, or after the allow-list was
- * tightened — we re-sanitize from the raw columns and persist the result
- * (fire-and-forget) so subsequent reads are fast again.
- *
- * The content (`content_*`) and full-content (`full_content_*`) families are
- * versioned independently because they are written at different times.
- */
-async function resolveSanitizedContent(
-  db: typeof import("@/server/db").db,
-  entryId: string,
-  stored: {
-    contentOriginalSanitized: string | null;
-    contentCleanedSanitized: string | null;
-    contentSanitizedVersion: number | null;
-    fullContentOriginalSanitized: string | null;
-    fullContentCleanedSanitized: string | null;
-    fullContentSanitizedVersion: number | null;
-  }
-) {
-  const contentCurrent = stored.contentSanitizedVersion === SANITIZER_VERSION;
-  const fullContentCurrent = stored.fullContentSanitizedVersion === SANITIZER_VERSION;
-
-  // Fast path: both families already sanitized at the current version.
-  if (contentCurrent && fullContentCurrent) {
-    return {
-      contentOriginal: stored.contentOriginalSanitized,
-      contentCleaned: stored.contentCleanedSanitized,
-      fullContentOriginal: stored.fullContentOriginalSanitized,
-      fullContentCleaned: stored.fullContentCleanedSanitized,
-    };
-  }
-
-  // Heal path: fetch the raw columns for whichever family is stale, sanitize,
-  // and persist so we don't pay this again.
-  const [raw] = await db
-    .select({
-      contentOriginal: entries.contentOriginal,
-      contentCleaned: entries.contentCleaned,
-      fullContentOriginal: entries.fullContentOriginal,
-      fullContentCleaned: entries.fullContentCleaned,
-    })
-    .from(entries)
-    .where(eq(entries.id, entryId))
-    .limit(1);
-
-  const resolved = {
-    contentOriginal: contentCurrent
-      ? stored.contentOriginalSanitized
-      : sanitizeEntryHtml(raw?.contentOriginal ?? null),
-    contentCleaned: contentCurrent
-      ? stored.contentCleanedSanitized
-      : sanitizeEntryHtml(raw?.contentCleaned ?? null),
-    fullContentOriginal: fullContentCurrent
-      ? stored.fullContentOriginalSanitized
-      : sanitizeEntryHtml(raw?.fullContentOriginal ?? null),
-    fullContentCleaned: fullContentCurrent
-      ? stored.fullContentCleanedSanitized
-      : sanitizeEntryHtml(raw?.fullContentCleaned ?? null),
-  };
-
-  // Persist the healed columns (fire-and-forget) so we don't pay this again.
-  // Each family is written in its own UPDATE gated on its version still being
-  // stale (`IS DISTINCT FROM`, which also matches NULL). This is a
-  // compare-and-swap: if a concurrent writer (e.g. a feed refresh via
-  // updateEntryContent, or a full-content fetch) has already stored fresh raw +
-  // sanitized + current version, our predicate is false and we skip — so a
-  // backfill computed from now-stale raw can never clobber newer content. The
-  // families are split because they version and change independently.
-  const heals: Promise<unknown>[] = [];
-  if (!contentCurrent) {
-    heals.push(
-      db
-        .update(entries)
-        .set({
-          contentOriginalSanitized: resolved.contentOriginal,
-          contentCleanedSanitized: resolved.contentCleaned,
-          contentSanitizedVersion: SANITIZER_VERSION,
-        })
-        .where(
-          and(
-            eq(entries.id, entryId),
-            sql`${entries.contentSanitizedVersion} IS DISTINCT FROM ${SANITIZER_VERSION}`
-          )
-        )
-    );
-  }
-  if (!fullContentCurrent) {
-    heals.push(
-      db
-        .update(entries)
-        .set({
-          fullContentOriginalSanitized: resolved.fullContentOriginal,
-          fullContentCleanedSanitized: resolved.fullContentCleaned,
-          fullContentSanitizedVersion: SANITIZER_VERSION,
-        })
-        .where(
-          and(
-            eq(entries.id, entryId),
-            sql`${entries.fullContentSanitizedVersion} IS DISTINCT FROM ${SANITIZER_VERSION}`
-          )
-        )
-    );
-  }
-  // A failed backfill write must not fail the read.
-  void (async () => {
-    try {
-      await Promise.all(heals);
-    } catch (err) {
-      logger.warn("Failed to persist re-sanitized entry content", { entryId, err });
-    }
-  })();
-
-  return resolved;
-}
-
-/**
- * Transform a raw full entry row into the entryFullSchema shape.
- * Strips internal fields, resolves sanitized content, and defaults fetchFullContent.
- */
-async function toFullEntry(
-  db: typeof import("@/server/db").db,
-  row: NonNullable<Awaited<ReturnType<typeof selectFullEntry>>>
-) {
-  const {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    hasStarred,
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    hasMarkedUnread,
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    hasMarkedReadOnList,
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    contentHash,
-    contentOriginalSanitized,
-    contentCleanedSanitized,
-    contentSanitizedVersion,
-    fullContentOriginalSanitized,
-    fullContentCleanedSanitized,
-    fullContentSanitizedVersion,
-    ...rest
-  } = row;
-
-  const content = await resolveSanitizedContent(db, row.id, {
-    contentOriginalSanitized,
-    contentCleanedSanitized,
-    contentSanitizedVersion,
-    fullContentOriginalSanitized,
-    fullContentCleanedSanitized,
-    fullContentSanitizedVersion,
-  });
-
-  return {
-    ...rest,
-    contentOriginal: content.contentOriginal,
-    contentCleaned: content.contentCleaned,
-    fullContentOriginal: content.fullContentOriginal,
-    fullContentCleaned: content.fullContentCleaned,
-    fetchFullContent: row.fetchFullContent ?? false,
-  };
-}
-
-// ============================================================================
 // Router
 // ============================================================================
+//
+// Full-entry reads go through `entriesService.selectFullEntry` /
+// `entriesService.toFullEntry`, which resolve the persisted sanitized content
+// (the read-path sanitization chokepoint lives in the services layer so MCP,
+// Google Reader, and Wallabag get the same guarantee).
 
 export const entriesRouter = createTRPCRouter({
   /**
@@ -552,12 +317,12 @@ export const entriesRouter = createTRPCRouter({
     .query(async ({ ctx, input }) => {
       const userId = ctx.session.user.id;
 
-      const row = await selectFullEntry(ctx.db, userId, input.id);
+      const row = await entriesService.selectFullEntry(ctx.db, userId, input.id);
       if (!row) {
         throw errors.entryNotFound();
       }
 
-      return { entry: await toFullEntry(ctx.db, row) };
+      return { entry: await entriesService.toFullEntry(ctx.db, row) };
     }),
 
   /**
@@ -871,14 +636,14 @@ export const entriesRouter = createTRPCRouter({
       const userId = ctx.session.user.id;
 
       // First, verify the entry exists and user has access
-      const rawEntry = await selectFullEntry(ctx.db, userId, input.id);
+      const rawEntry = await entriesService.selectFullEntry(ctx.db, userId, input.id);
 
       if (!rawEntry) {
         throw errors.entryNotFound();
       }
 
       const entry = {
-        ...(await toFullEntry(ctx.db, rawEntry)),
+        ...(await entriesService.toFullEntry(ctx.db, rawEntry)),
         contentHash: rawEntry.contentHash,
       };
